renamed and added docs for mkdocs
