+++ docs/utils/fps.md
@@ -1,0 +1,3 @@
+import supervision.utils.fps
+
+:::supervision.utils.fps.FpsMonitor
