+++ mkdocs.yml
@@ -48,6 +48,7 @@ nav:
         - Image: utils/image.md
         - Notebook: utils/notebook.md
         - File: utils/file.md
+        - Fps: utils/fps.md
   - Changelog: changelog.md
 
 theme:
