+++ supervision/__init__.py
@@ -34,7 +34,7 @@ from supervision.metrics.detection import ConfusionMatrix
 from supervision.utils.file import list_files_with_extensions
 from supervision.utils.image import ImageSink, crop
 from supervision.utils.notebook import plot_image, plot_images_grid
-from supervision.utils.fpsmonitor import FpsMonitor
+from supervision.utils.fps import FpsMonitor
 from supervision.utils.video import (
     VideoInfo,
     VideoSink,
